se realizo las pruebas unitarias de clientes y cuentas junto algunas mejoras de logica y orden de archivos

+++ src/main/java/com/prueba/demo/repository/CuentaRepositorio.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 
 public interface CuentaRepositorio extends JpaRepository<Cuenta, Long> {
-    boolean existsByNumeroCuenta(Integer numeroCuenta);
+    boolean existsByNumeroCuenta(Long numeroCuenta);
 
     List<Cuenta> findByCliente(Cliente cliente);
 }
